Added OMOP methods to shared_code dir
